// src/app/api/model/ratedTvEpisodeListItem.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RatedTvEpisodeListItem { 
    air_date?: string;
    episode_number?: number;
    id?: number;
    name?: string;
    overview?: string;
    production_code?: string;
    runtime?: number | null;
    season_number?: number;
    show_id?: number;
    still_path?: string | null;
    vote_average?: number;
    vote_count?: number;
    rating?: number;
}