// src/app/api/model/ratedTvSeriesListItem.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RatedTvSeriesListItem { 
    adult?: boolean;
    backdrop_path?: string | null;
    genre_ids?: Array<number>;
    id?: number;
    origin_country?: Array<string>;
    original_language?: string;
    original_name?: string;
    overview?: string;
    popularity?: number;
    poster_path?: string | null;
    first_air_date?: string;
    name?: string;
    vote_average?: number;
    vote_count?: number;
    rating?: number;
}